Refactor integration tests and add account test suite
Moved account-related integration tests from main.cpp to a new account_repository_db_it.cpp under tests/integration/account. Added supporting CMakeLists.txt and test_database.hpp for test database setup. Updated tests/integration/CMakeLists.txt to include the new account test directory. main.cpp now contains only an empty main function.
