+++ tests/integration/CMakeLists.txt
@@ -1,1 +1,4 @@
 setup_test(canary_it integration)
+
+add_subdirectory(account)
+
