+++ tests/integration/account/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(canary_it PRIVATE
+    account_repository_db_it.cpp
+)
+
